+++ tests/src/FunctionalJavascript/Integration/RiddleTest.php
@@ -72,7 +72,7 @@ class RiddleTest extends ThunderJavascriptTestBase {
 
     // Click button to load Riddles and compare thumbnails.
     $this->clickButtonDrupalSelector($page, 'edit-import-riddle');
-    $this->assertNotEmpty($this->assertSession()->waitForElementVisible('css', '.view-media-entity-browser .views-field-thumbnail__target-id img'));
+    $this->assertNotEmpty($this->assertSession()->waitForElementVisible('css', '.view-media-entity-browser .views-field-thumbnail__target-id img:last-child'));
 
     // Close entity browser.
     $this->getSession()->switchToIFrame();