+++ tests/src/FunctionalJavascript/MediaImageModifyTest.php
@@ -14,6 +14,9 @@ use Imagick;
  */
 class MediaImageModifyTest extends ThunderJavascriptTestBase {
 
+  use ThunderEntityBrowserTestTrait;
+  use ThunderParagraphsTestTrait;
+
   /**
    * Test Focal Point change.
    */
@@ -103,6 +106,49 @@ class MediaImageModifyTest extends ThunderJavascriptTestBase {
       ->fieldValueEquals('field_source[0][value]', "Media {$mediaId} Source");
     $this->assertSession()
       ->fieldValueEquals('field_description[0][value]', "<p>Media {$mediaId} Description</p>");
+  }
+
+  /**
+   * Test add/remove image in image paragraph.
+   *
+   * Demo Article (node Id: 6) is used for testing.
+   * Cases tested:
+   *   - remove inside inline entity form
+   *   - add inside entity browser
+   *   - reorder inside entity browser
+   *   - remove inside entity browser.
+   */
+  public function testRemoveAdd() {
+
+    // Test remove inside inline entity form.
+    $this->drupalGet("node/6/edit");
+
+    $page = $this->getSession()->getPage();
+
+    $this->editParagraph($page, 'field_paragraphs', 0);
+
+    // Remove image.
+    $this->clickAjaxButtonCssSelector('[data-drupal-selector="edit-field-paragraphs-0-subform-field-image-current-items-0-remove-button"]');
+    $this->assertSession()->assertWaitOnAjaxRequest();
+
+    // Check that there are no errors.
+    $this->assertSession()
+      ->elementNotExists('css', '[data-drupal-selector="edit-field-paragraphs-0-subform-field-image-wrapper"] div.messages--error');
+
+    // Click Select entities -> to open Entity Browser.
+    $this->openEntityBrowser($page, 'edit-field-paragraphs-0-subform-field-image-entity-browser-entity-browser-open-modal', 'image_browser');
+
+    // Select another image and store filename.
+    $this->clickButtonCssSelector($page, '#entity-browser-image-browser-form div.view-content > div.views-row:nth-child(1)');
+    $fileName = $this->getSession()->evaluateScript('jQuery(\'#entity-browser-image-browser-form div.view-content > div.views-row:nth-child(1) img\').attr(\'src\').split(\'?\')[0].split(\'/\').splice(-1);');
+    $this->clickButtonDrupalSelector($page, 'edit-submit');
+    $this->getSession()->switchToIFrame();
+    $this->assertSession()->assertWaitOnAjaxRequest();
+
+    // Save paragraph.
+    $this->clickAjaxButtonCssSelector('[name="field_paragraphs_0_collapse"]');
+
+    $this->assertEquals($fileName, $this->getSession()->evaluateScript('jQuery(\'[data-drupal-selector="edit-field-paragraphs-0-preview"] div.paragraph-preview__thumbnail img\').attr(\'src\').split(\'?\')[0].split(\'/\').splice(-1)'), 'Image file should be identical to previously selected.');
   }
 
 }