+++ tests/src/Nightwatch/Commands/performance/startMeasurement.js
@@ -19,29 +19,32 @@
 
 exports.command = function startMeasurement(
   serverUrl,
-  serviceName,
   transactionName,
   domain
 ) {
   const browser = this;
 
   browser.perform(() => {
-    const apmInstance = browser.apm.start({
-      serverUrl,
-      serviceName
-    });
+    if (!browser.apm.isStarted()) {
+      browser.apm.start({ serverUrl, serviceName: "NightwatchJS - Test" });
+    }
 
-    browser.apmDomain = domain;
-    browser.apmTrans = apmInstance.startTransaction(transactionName, "test");
-    browser.apmSpans = [];
+    browser.globals.apmDomain = domain;
+    browser.globals.apmTrans = browser.apm.startTransaction(
+      transactionName,
+      "test"
+    );
+    browser.globals.apmSpans = [];
 
     browser
+      // We need to open some URL before set cookie.
+      .drupalRelativeURL("/")
       .setCookie({
         domain,
         httpOnly: false,
         path: "/",
         name: "traceId",
-        value: browser.apmTrans.traceId
+        value: browser.globals.apmTrans.traceId
       })
       .setCookie({
         domain,
@@ -49,7 +52,11 @@ exports.command = function startMeasurement(
         path: "/",
         name: "serverUrl",
         value: serverUrl
-      })
+      });
+
+    // Label set on Node.JS APM agent should be also set for Browser APM agent.
+    browser.performance
+      .setLabel("branch", process.env.THUNDER_BRANCH)
       .setCookie({
         domain,
         httpOnly: false,
@@ -57,7 +64,14 @@ exports.command = function startMeasurement(
         name: "branchTag",
         value: process.env.THUNDER_BRANCH
       })
-      .performance.setLabel("branch", process.env.THUNDER_BRANCH);
+      .performance.setLabel("test", browser.currentTest.name)
+      .setCookie({
+        domain,
+        httpOnly: false,
+        path: "/",
+        name: "testTag",
+        value: browser.currentTest.name
+      });
   });
 
   return browser;