+++ tests/src/Nightwatch/Tests/ArticleCreate.js
@@ -24,10 +24,8 @@ module.exports = {
   createAnArticleWithParagraphs(browser) {
     browser
       .resizeWindow(1024, 1024)
-      .drupalRelativeURL("/user/login")
       .performance.startMeasurement(
         process.env.THUNDER_APM_URL,
-        "NightwatchJS - Test",
         "Create an article with paragraphs",
         `.${process.env.THUNDER_SITE_HOSTNAME}`
       )